Refactor codebase structure: flatten ast/codegen/errors, merge semantic models
- Flatten `src/ast/mod.rs` to `src/ast.rs`
- Flatten `src/codegen/mod.rs` to `src/codegen.rs`
- Consolidate `src/errors/` into `src/errors.rs` (merging `assembly` and `messages`)
- Merge `src/semantic/assembly_model.rs` into `src/semantic/assembler.rs`
- Move `match_suffix` from `src/morphology/matcher.rs` to `src/morphology/mod.rs`
- Verify all tests pass and fix imports
