+++ src/semantic/assembler_tests.rs
@@ -2,9 +2,9 @@ use crate::ast::{Expr, Word};
 use crate::morphology::{Case, Gender, MorphAnalysis, Number, PartOfSpeech, Person, analyze};
 use crate::semantic::AssemblyError;
 use crate::semantic::assembler::{
-    Assembler, MAX_ADJECTIVES, MAX_ARRAYS, MAX_BLOCKS, MAX_GENITIVES, MAX_INDEX_ACCESSES,
-    MAX_LITERALS, MAX_NESTED_PHRASES, MAX_NOMINATIVES, MAX_OPERATORS, MAX_PARTICIPLES,
-    MAX_PROPERTY_ACCESSES, MAX_UNWRAPS,
+    Assembler, AssembledStatement, Constituent, MAX_ADJECTIVES, MAX_ARRAYS, MAX_BLOCKS,
+    MAX_GENITIVES, MAX_INDEX_ACCESSES, MAX_LITERALS, MAX_NESTED_PHRASES, MAX_NOMINATIVES,
+    MAX_OPERATORS, MAX_PARTICIPLES, MAX_PROPERTY_ACCESSES, MAX_UNWRAPS,
 };
 use std::borrow::Cow;
 
@@ -796,7 +796,6 @@ fn test_verbless_statement() {
 
 #[test]
 fn test_assembled_statement_derive_coverage() {
-    use crate::semantic::assembly_model::AssembledStatement;
     // Cover #[derive(Clone, Debug, Default)] for AssembledStatement
     let stmt = AssembledStatement::default();
     let cloned = stmt.clone();
@@ -978,7 +977,6 @@ fn test_assembler_error_cases_coverage() {
 #[test]
 fn test_constituent_derive_coverage() {
     use crate::morphology::Case;
-    use crate::semantic::assembly_model::Constituent;
     use smol_str::SmolStr;
 
     let c = Constituent {
